Revert "CI: run on pull requests"
This reverts commit 666a4643f908713f1a52fe66087af35cbd8f501c.

It turned out to be a bit abusive, since Github doesn't deduplicate
such actions; every time I pushed to a PR-eligible branch, it would
build the whole project TWICE. That's just a waste of resources.

We need to get some kind of test suite which is run before the sdist
action is deemed a success, and run *that* on each commit; then we
could consider not running the full bdist suite on every commit, but
only on PRs?